+++ Support/Plugins/Tester/Tester.cpp
@@ -109,13 +109,13 @@ bool PositionInteract::draw(const OFX::DrawArgs &args)
 // overridden functions from OFX::Interact to do things
 bool PositionInteract::penMotion(const OFX::PenArgs &args)
 {
-  // figure the size of the box in cannonical coords
+  // figure the size of the box in canonical coords
   float dx = (float)(kBoxSize.x / args.pixelScale.x);
   float dy = (float)(kBoxSize.y / args.pixelScale.y);
 
   OfxPointD pos = getCanonicalPosition(args.time);
 
-  // pen position is in cannonical coords
+  // pen position is in canonical coords
   OfxPointD penPos = args.penPosition;
 
   switch(_state)